Project 3 - Image to Recipe Generator/requirements.txt

streamlit==1.35.0
transformers==4.41.1
langchain==0.2.0
langchain-google-genai==0.0.8
python-dotenv==1.0.1
requests==2.31.0
